Move identifier metadata normalization from utils into dedicated directory
